requirements.txt
Package                       Version
----------------------------- --------
asgiref                       3.9.1
attrs                         25.3.0
dj-rest-auth                  7.0.1
Django                        5.2.4
django-allauth                65.10.0
django-cors-headers           4.7.0
django-jazzmin                3.0.1
djangorestframework           3.16.0
djangorestframework_simplejwt 5.5.0
drf-spectacular               0.28.0
inflection                    0.5.1
jsonschema                    4.25.0
jsonschema-specifications     2025.4.1
pip                           25.1.1
psycopg2-binary               2.9.10
PyJWT                         2.9.0
PyYAML                        6.0.2
referencing                   0.36.2
rpds-py                       0.26.0
setuptools                    80.9.0
sqlparse                      0.5.3
typing_extensions             4.14.1
tzdata                        2025.2
uritemplate                   4.2.0
wheel                         0.45.1
